Refactor: organize ui folder to be more practical
- Move main views to their own folder
- Move list items together. They already share code and will
even more in the future.
- Refactor the itinerary code to be a bit more logical. There
is now a form, a query, and a result.
- move form_components to components.
