+++ lib/main.dart
@@ -2,10 +2,10 @@ import 'package:bussit/api/hsl_api.dart';
 import 'package:bussit/database/database.dart';
 import 'package:bussit/model/saved_stops.dart';
 import 'package:bussit/model/user_actions.dart';
-import 'package:bussit/ui/itinerary_view.dart';
+import 'package:bussit/ui/views/itinerary_view.dart';
 import 'package:bussit/ui/widgets/map/map_widget.dart';
-import 'package:bussit/ui/saved_stops_view.dart';
-import 'package:bussit/ui/stop_search_view.dart';
+import 'package:bussit/ui/views/saved_stops_view.dart';
+import 'package:bussit/ui/views/stop_search_view.dart';
 import 'package:flutter/material.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
 import 'package:provider/provider.dart';
