+++ lib/ui/views/saved_stops_view.dart
@@ -1,6 +1,6 @@
 import 'package:bussit/model/saved_stops.dart';
 import 'package:flutter/material.dart';
-import 'package:bussit/ui/widgets/stops/stop_list.dart';
+import 'package:bussit/ui/widgets/lists/stop_list.dart';
 import 'package:provider/provider.dart';
 import 'dart:developer' as developer;
