+++ lib/ui/widgets/itinerary/itinerary_query.dart
@@ -1,17 +1,19 @@
 import 'package:bussit/graphql/itinerary_query.graphql.dart';
 import 'package:bussit/graphql/schema.graphql.dart';
+import 'package:bussit/graphql/trip_query.graphql.dart';
 import 'package:bussit/model/address.dart';
-import 'package:bussit/ui/widgets/itineraries/itinerary_item.dart';
-import 'package:flutter/material.dart';
+import 'package:bussit/ui/widgets/itinerary/itinerary_results.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:graphql_flutter/graphql_flutter.dart';
-import 'package:bussit/utils/graphql_hooks.dart';
-import 'dart:developer' as developer;
+import 'package:graphql/client.dart';
 import 'package:intl/intl.dart';
-import '../../../graphql/trip_query.graphql.dart';
+import 'dart:developer' as developer;
 
-class ItineraryVariables extends HookWidget {
-  ItineraryVariables({
+import '../../../utils/graphql_hooks.dart';
+
+class ItineraryQuery extends HookWidget {
+  /// Store the parameters extracted from an ItineraryForm
+  ItineraryQuery({
     required this.from,
     required this.to,
     this.nResults,
@@ -63,48 +65,6 @@ class ItineraryVariables extends HookWidget {
     variables = useItineraryVariables(variables, from, to, time);
     return ItineraryResults(variables: variables);
   }
-}
-
-/// Widget showing a list of stops
-class ItineraryResults extends HookWidget {
-  const ItineraryResults({required this.variables, super.key});
-  final Variables$Query$Itinerary variables;
-  @override
-  Widget build(BuildContext context) {
-    final result = useQueryLifecycleAware(Options$Query$Itinerary(
-      fetchPolicy: FetchPolicy.networkOnly,
-      variables: variables,
-    ));
-
-    return itineraryListBuilder(result.result);
-  }
-}
-
-// Build a stop list from a query result
-Widget itineraryListBuilder(QueryResult? result,
-    {VoidCallback? refetch, FetchMore? fetchMore}) {
-  if (result == null) {
-    return const Text("No result...");
-  }
-  if (result.hasException) {
-    return Text(result.exception.toString());
-  }
-
-  if (result.isLoading) {
-    return const Text('Loading...');
-  }
-  List<Query$Itinerary$plan$itineraries?>? list;
-
-  if (result.data != null) {
-    list = Query$Itinerary.fromJson(result.data!).plan?.itineraries;
-  } else {
-    return const Text("List is null :(");
-  }
-  list ??= [];
-
-  return Column(
-    children: list.map((e) => ItineraryWidget(itinerary: e)).toList(),
-  );
 }
 
 /// A hook to get relevant variables from start and end locations
