+++ lib/ui/widgets/lists/stop_list.dart
@@ -1,5 +1,5 @@
 import 'package:bussit/graphql/stops_query.graphql.dart';
-import 'package:bussit/ui/widgets/stops/stop_item.dart';
+import 'package:bussit/ui/widgets/list_items/stop_item.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
